[development/directx_init_drawing] Add some additional documentation Move none templated functions to the source file
